[Refactor] Extract module MinBodyLength
Two methods that are exactly the same in two classes.
